pyproject.toml
[tool.poetry]
name = "sky-alert"
version = "0.1.0"
description = "Alert that combines information from multiple data sources for what could be a potentially good stargazing / astrophotography night opportunity for your locale."
authors = ["Rohan Sahgal <rohansahgal@hotmail.com>", "Chris Zhu <thereisnoaddress@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.105.0"
uvicorn = "^0.24.0.post1"
python-dotenv = "^1.0.0"
pre-commit = "^3.6.0"
mypy = "^1.7.1"
flake8 = "^6.1.0"
requests = "^2.31.0"
types-requests = "^2.31.0"
pytest = "^7.4.3"
httpx = "^0.26.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''
